requirements.txt
alibabacloud-dingtalk==1.0.0
alibabacloud-tea-openapi==0.3.0
alibabacloud-tea-util==0.2.0
requests==2.26.0
beautifulsoup4==4.10.0
python-dotenv==0.19.0